+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### next
 - Unreleased
+- Change(tui): increase connection timeout from 5s to 30s, but print a message on original time.
 - Fix(server): on linux+mpris, set volume on start instead of only on change.
 
 ### [V0.11.0]